transfer tfw_current_timestamp() in the library
